mu2eprodsys: add backward compatibility for older Offline releases.
Offline versions using older FHICLCPP need to use fhicl-getpar from
mu2ebintools, while versions based on art v3_08_00 and newer do not
need the extra package.  The worker node script simply uses the
"fhicl-get" command.  The logic to provide the correct implementation
for that command based on FHICLCPP version is in the mu2etools
package, which from a "sometimes" dependency changed to something that
we need to set up in every job.  Add a new --mu2etools-version
mu2eprodsys option that goes with that change.